requirements.txt
torch==2.1.0
torchvision==0.16.0

# Tokenization
sentencepiece==0.1.99

# Dataset and utilities
datasets==2.17.1
Pillow==10.0.1
numpy==1.24.3
tqdm==4.66.1
scikit-learn==1.3.0
matplotlib==3.7.2